web-version/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.10.6
python-dotenv==1.0.0
requests==2.31.0
Crawl4AI==0.4.247
python-multipart==0.0.6
aiohttp==3.9.1
httpx==0.25.2